+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - refactor: generalize HTTP listener config loading
 - refactor: generalize HTTP listener creation
+- feat!: add client log collection as a tsnet listener
 
 ## 0.0.2 (2024-02-22)